fix bug

+++ OurNet/MyTraining.py
@@ -79,7 +79,7 @@ for epoch in range(opt.nepoch):
         points, target = data
         points1 = points[0].transpose(2, 1)
         points2 = points[1].transpose(2, 1)
-        target = torch.tensor(target, dtype=torch.long)
+        target = torch.tensor(target, dtype=torch.float32)
         points1 = points1.type(torch.FloatTensor)
         points2 = points2.type(torch.FloatTensor)
         points1, points2, target = points1.cuda(), points2.cuda(), target.cuda()
@@ -87,7 +87,7 @@ for epoch in range(opt.nepoch):
         classifier = classifier.train()
         pred1, pred2 = classifier(points1, points2)
         pred1 = pred1.unsqueeze(0)
-        target_cls = target[4].unsqueeze(0)
+        target_cls = target[4].to(torch.float32).unsqueeze(0)
         loss1 = F.cross_entropy(pred1, target_cls).to(torch.float32)
         loss2 = F.mse_loss(pred2, target[:4].unsqueeze(0).to(torch.float32))
         # if the actual value of target[4] is 0, then the loss2 is 0
@@ -103,7 +103,7 @@ for epoch in range(opt.nepoch):
             points, target = data
             points1 = points[0].transpose(2, 1)
             points2 = points[1].transpose(2, 1)
-            target = torch.tensor(target, dtype=torch.long)
+            target = torch.tensor(target, dtype=torch.float32)
             points1 = points1.type(torch.FloatTensor)
             points2 = points2.type(torch.FloatTensor)
             points1, points2, target = points1.cuda(), points2.cuda(), target.cuda()
@@ -111,7 +111,7 @@ for epoch in range(opt.nepoch):
             classifier = classifier.train()
             pred1, pred2 = classifier(points1, points2)
             pred1 = pred1.unsqueeze(0)
-            target_cls = target[4].unsqueeze(0)
+            target_cls = target[4].to(torch.float32).unsqueeze(0)
             loss1 = F.cross_entropy(pred1, target_cls).to(torch.float32)
             loss2 = F.mse_loss(pred2, target[:4].unsqueeze(0).to(torch.float32))
             loss2 = loss2 * (target[4] != 0).float()
